separate update user and delete user
